+++ instana-example/app/src/main/java/com/instana/mobileeum/ui/custom/CustomEventFragment.kt
@@ -16,10 +16,9 @@ import android.widget.ArrayAdapter
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.instana.mobileeum.R
+import com.instana.mobileeum.databinding.FragmentCustomBinding
 import com.instana.mobileeum.notBlankOrNull
 import com.instana.mobileeum.toMap
-import kotlinx.android.synthetic.main.fragment_custom.view.*
 
 class CustomEventFragment : Fragment() {
 
@@ -30,66 +29,76 @@ class CustomEventFragment : Fragment() {
 
     private var textChangeListener: TextWatcher? = null
 
+    private var _binding: FragmentCustomBinding? = null
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         viewModel = ViewModelProvider(this).get(CustomEventViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_custom, container, false)
+        _binding = FragmentCustomBinding.inflate(inflater, container, false)
+        val root = binding.root
 
         // Request Presets
         viewModel.presets.observe(viewLifecycleOwner, Observer { list ->
             presets = list
             val spinnerOptions = list.map { it.presetName }.toMutableList().apply { add("CUSTOM") }
-            root.eventPresets.adapter = ArrayAdapter(requireContext(), android.R.layout.simple_list_item_1, spinnerOptions)
+            binding.eventPresets.adapter = ArrayAdapter(requireContext(), android.R.layout.simple_list_item_1, spinnerOptions)
         })
-        root.eventPresets.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+        binding.eventPresets.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
             override fun onNothingSelected(p0: AdapterView<*>?) {}
             override fun onItemSelected(p0: AdapterView<*>?, p1: View?, position: Int, p3: Long) {
                 presets.getOrNull(position)?.run {
-                    listenEditTextChanges(root, false)
-                    root.eventName.setText(this.eventName)
-                    root.eventStartTime.setText(this.startTime?.toString())
-                    root.eventDuration.setText(this.duration?.toString())
-                    root.eventViewName.setText(this.viewName)
-                    root.eventMeta.setText(this.meta?.toString())
-                    root.eventBackendTracingId.setText(this.backendTracingId)
-                    root.eventError.setText(this.error)
-                    listenEditTextChanges(root, true)
+                    listenEditTextChanges(false)
+                    binding.eventName.setText(this.eventName)
+                    binding.eventStartTime.setText(this.startTime?.toString())
+                    binding.eventDuration.setText(this.duration?.toString())
+                    binding.eventViewName.setText(this.viewName)
+                    binding.eventMeta.setText(this.meta?.toString())
+                    binding.eventBackendTracingId.setText(this.backendTracingId)
+                    binding.eventError.setText(this.error)
+                    listenEditTextChanges(true)
                 }
             }
         }
 
         // EditTexts
-        listenEditTextChanges(root, true)
+        listenEditTextChanges(true)
 
         // Run Button
-        root.sendEvent.setOnClickListener {
+        binding.sendEvent.setOnClickListener {
             waitingResponse = true
             viewModel.sendCustomEvent(
-                eventName = root.eventName.text.toString(),
-                startTimeEpochMs = root.eventStartTime.text?.toString()?.notBlankOrNull()?.toLong(),
-                durationMs = root.eventDuration.text?.toString()?.notBlankOrNull()?.toLong(),
-                viewName = root.eventViewName.text?.toString()?.notBlankOrNull(),
-                meta = root.eventMeta.text?.toString()?.notBlankOrNull()?.toMap(),
-                backendTracingID = root.eventBackendTracingId.text?.toString()?.notBlankOrNull(),
-                errorMessage = root.eventError.text?.toString()?.notBlankOrNull()
+                eventName = binding.eventName.text.toString(),
+                startTimeEpochMs = binding.eventStartTime.text?.toString()?.notBlankOrNull()?.toLong(),
+                durationMs = binding.eventDuration.text?.toString()?.notBlankOrNull()?.toLong(),
+                viewName = binding.eventViewName.text?.toString()?.notBlankOrNull(),
+                meta = binding.eventMeta.text?.toString()?.notBlankOrNull()?.toMap(),
+                backendTracingID = binding.eventBackendTracingId.text?.toString()?.notBlankOrNull(),
+                errorMessage = binding.eventError.text?.toString()?.notBlankOrNull()
             )
         }
 
         return root
     }
 
-    private fun listenEditTextChanges(root: View, enabled: Boolean) {
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    private fun listenEditTextChanges(enabled: Boolean) {
         val textInputs = listOf(
-            root.eventName,
-            root.eventStartTime,
-            root.eventDuration,
-            root.eventViewName,
-            root.eventMeta,
-            root.eventBackendTracingId,
-            root.eventError
+            binding.eventName,
+            binding.eventStartTime,
+            binding.eventDuration,
+            binding.eventViewName,
+            binding.eventMeta,
+            binding.eventBackendTracingId,
+            binding.eventError
         )
         textChangeListener?.run {
             textInputs.forEach { it.removeTextChangedListener(this) }
@@ -101,7 +110,7 @@ class CustomEventFragment : Fragment() {
                 override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
                     val textInputFocused = textInputs.any { it.isFocused }
                     if (textInputFocused) {
-                        root.eventPresets.setSelection(presets.size)
+                        binding.eventPresets.setSelection(presets.size)
                     }
                 }
             }