+++ instana-example/app/src/main/java/com/instana/mobileeum/ui/httpurlconnection/HttpURLConnectionFragment.kt
@@ -16,9 +16,8 @@ import android.widget.ArrayAdapter
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.instana.mobileeum.R
+import com.instana.mobileeum.databinding.FragmentHttpurlconnectionBinding
 import com.instana.mobileeum.network.RequestPreset
-import kotlinx.android.synthetic.main.fragment_httpurlconnection.view.*
 
 class HttpURLConnectionFragment : Fragment() {
 
@@ -29,74 +28,84 @@ class HttpURLConnectionFragment : Fragment() {
 
     private var textChangeListener: TextWatcher? = null
 
+    private var _binding: FragmentHttpurlconnectionBinding? = null
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         viewModel = ViewModelProvider(this).get(HttpURLConnectionViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_httpurlconnection, container, false)
+        _binding = FragmentHttpurlconnectionBinding.inflate(inflater, container, false)
+        val root = binding.root
 
         // Request Presets
         viewModel.presets.observe(viewLifecycleOwner, Observer { list ->
             presets = list
             val spinnerOptions = list.map { it.name }.toMutableList().apply { add("CUSTOM") }
-            root.requestPresets.adapter = ArrayAdapter(requireContext(), android.R.layout.simple_list_item_1, spinnerOptions)
+            binding.requestPresets.adapter = ArrayAdapter(requireContext(), android.R.layout.simple_list_item_1, spinnerOptions)
         })
-        root.requestPresets.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+        binding.requestPresets.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
             override fun onNothingSelected(p0: AdapterView<*>?) {}
             override fun onItemSelected(p0: AdapterView<*>?, p1: View?, position: Int, p3: Long) {
                 presets.getOrNull(position)?.run {
-                    listenEditTextChanges(root, false)
-                    root.requestMethod.setText(this.method)
-                    root.requestUrl.setText(this.url)
-                    root.requestBody.setText(this.body)
-                    listenEditTextChanges(root, true)
+                    listenEditTextChanges(false)
+                    binding.requestMethod.setText(this.method)
+                    binding.requestUrl.setText(this.url)
+                    binding.requestBody.setText(this.body)
+                    listenEditTextChanges(true)
                 }
             }
         }
 
         // EditTexts
-        listenEditTextChanges(root, true)
+        listenEditTextChanges(true)
 
         // Run Button
-        root.runRequest.setOnClickListener {
+        binding.runRequest.setOnClickListener {
             waitingResponse = true
             viewModel.executeRequest(
-                method = root.requestMethod.text.toString(),
-                url = root.requestUrl.text.toString(),
-                body = if (root.requestBody.text.isNullOrBlank()) null else root.requestBody.text.toString()
+                method = binding.requestMethod.text.toString(),
+                url = binding.requestUrl.text.toString(),
+                body = if (binding.requestBody.text.isNullOrBlank()) null else binding.requestBody.text.toString()
             )
         }
 
         // Response
         viewModel.response.observe(viewLifecycleOwner, Observer {
-            root.response.text = it
-            if (waitingResponse) root.scrollView.fullScroll(View.FOCUS_DOWN)
+            binding.response.text = it
+            if (waitingResponse) binding.scrollView.fullScroll(View.FOCUS_DOWN)
         })
 
         return root
     }
 
-    private fun listenEditTextChanges(root: View, enabled: Boolean) {
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    private fun listenEditTextChanges(enabled: Boolean) {
         textChangeListener?.run {
-            root.requestMethod.removeTextChangedListener(this)
-            root.requestUrl.removeTextChangedListener(this)
-            root.requestBody.removeTextChangedListener(this)
+            binding.requestMethod.removeTextChangedListener(this)
+            binding.requestUrl.removeTextChangedListener(this)
+            binding.requestBody.removeTextChangedListener(this)
         }
         if (enabled) {
             textChangeListener = object : TextWatcher {
                 override fun afterTextChanged(p0: Editable?) {}
                 override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
                 override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
-                    if (root.requestMethod.isFocused || root.requestUrl.isFocused || root.requestBody.isFocused) {
-                        root.requestPresets.setSelection(presets.size)
+                    if (binding.requestMethod.isFocused || binding.requestUrl.isFocused || binding.requestBody.isFocused) {
+                        binding.requestPresets.setSelection(presets.size)
                     }
                 }
             }
-            root.requestMethod.addTextChangedListener(textChangeListener)
-            root.requestUrl.addTextChangedListener(textChangeListener)
-            root.requestBody.addTextChangedListener(textChangeListener)
+            binding.requestMethod.addTextChangedListener(textChangeListener)
+            binding.requestUrl.addTextChangedListener(textChangeListener)
+            binding.requestBody.addTextChangedListener(textChangeListener)
         }
     }
 }