+++ plugin/src/main/java/com/instana/android/plugin/transform/Transform.kt
@@ -5,76 +5,110 @@
 
 package com.instana.android.plugin.transform
 
-import com.android.build.api.transform.QualifiedContent
-import com.android.build.api.transform.Transform
-import com.android.build.api.transform.TransformInvocation
-import com.android.build.gradle.AppExtension
+import com.android.build.api.instrumentation.*
+import com.android.build.api.variant.AndroidComponentsExtension
 import com.instana.android.plugin.Extension
 import com.instana.android.plugin.IgnoredClasses
 import com.instana.android.plugin.logAll
+import com.instana.android.plugin.transform.asm.InstrumentationVisitor
+import com.instana.android.plugin.transform.asm.OkHttp3BuilderVisitor
 import org.gradle.api.Project
 import org.gradle.api.logging.Logging
+import org.gradle.api.provider.ListProperty
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
+import org.objectweb.asm.ClassVisitor
 
-class Transform(private val project: Project) : Transform() {
-
-    private val logger = Logging.getLogger(Transform::class.java)
-
-    override fun getName(): String {
-        return "InstanaTransform"
-    }
-
-    // Classes only (not resources)
-    private val typeClasses = setOf(QualifiedContent.DefaultContentType.CLASSES)
-
-    override fun getInputTypes(): Set<QualifiedContent.ContentType> {
-        return typeClasses
-    }
-
-
-    // Classes from all parts of the app
-    private val scopes = setOf(
-        QualifiedContent.Scope.PROJECT,
-        QualifiedContent.Scope.SUB_PROJECTS,
-        QualifiedContent.Scope.EXTERNAL_LIBRARIES
-    )
-
-    override fun getScopes(): MutableSet<in QualifiedContent.Scope> {
-        return scopes.toMutableSet()
-    }
-
-
-    // Handle incremental builds
-    override fun isIncremental(): Boolean {
-        return true
-    }
-
-
-    override fun transform(transformInvocation: TransformInvocation) {
+class Transform {
+    fun doIt(project: Project) {
         val start = System.currentTimeMillis()
 
         val ext = project.extensions.findByType(Extension::class.java) ?: Extension()
-        logger.logAll(ext)
 
-        val extIgnores = IgnoredClasses.from(ext.ignoreClassesRegex, logger)
-        val ignores = listOf(
-            IgnoredClasses.instanaLibraries,
-            *IgnoredClasses.troublesomeAnalytics.toTypedArray(),
-            *extIgnores.toTypedArray(),
-        )
+        var firstLoop = true
+        var ignores: List<Regex> = listOf()
 
-        val appExtension = project.extensions.findByName("android") as AppExtension
-        val config = TransformConfig(transformInvocation, appExtension.bootClasspath, ignores, ext)
+        val androidComponents = project.extensions.getByType(AndroidComponentsExtension::class.java)
+        androidComponents.onVariants { variant ->
+            if (firstLoop) {
+                firstLoop = false
+                project.logger.logAll(ext)
 
-        if (ext.enableWeaving) {
-            TransformImpl(config).doIt()
-        } else {
-            logger.warn("Skipping Instana transform")
-            TransformNoop(config).doIt()
+                val extIgnores = IgnoredClasses.from(ext.ignoreClassesRegex, project.logger)
+                ignores = listOf(
+                    IgnoredClasses.instanaLibraries,
+                    *IgnoredClasses.troublesomeAnalytics.toTypedArray(),
+                    *extIgnores.toTypedArray(),
+                )
+                project.logger.debug("Ignores list= ${ignores}")
+            }
+
+            variant.instrumentation.transformClassesWith(InstanaClassVisitorFactory::class.java,
+                InstrumentationScope.ALL) {
+                it.ignores.set(ignores)
+                it.logVisits.set(ext.logVisits)
+                it.logInstrumentation.set(ext.logInstrumentation)
+            }
+            variant.instrumentation.setAsmFramesComputationMode(FramesComputationMode.COMPUTE_FRAMES_FOR_INSTRUMENTED_METHODS)
         }
 
         val end = System.currentTimeMillis()
-        logger.info("Instana transform took ${end - start}ms")
-        if (ext.logTimeSpent) project.logger.warn("Instana transform took ${end - start}ms")
+        project.logger.info("Instana transform preparation took ${end - start}ms")
+        if (ext.logTimeSpent) project.logger.warn("Instana transform preparation took ${end - start}ms")
     }
 
+    interface InstanaInstruParams : InstrumentationParameters {
+        @get:Input
+        val ignores: ListProperty<Regex>
+
+        @get:Input
+        val logVisits: Property<Boolean>
+
+        @get:Input
+        val logInstrumentation: Property<Boolean>
+    }
+
+    abstract class InstanaClassVisitorFactory :
+        AsmClassVisitorFactory<InstanaInstruParams> {
+
+        override fun createClassVisitor(
+            classContext: ClassContext,
+            nextClassVisitor: ClassVisitor
+        ): ClassVisitor {
+
+            val config = InstrumentationConfig(parameters.get().logVisits.get(), parameters.get().logInstrumentation.get())
+            return when (classContext.currentClassData.className) {
+                "okhttp3.OkHttpClient\$Builder" -> {
+                    OkHttp3BuilderVisitor(nextClassVisitor, config)
+                }
+                else -> {
+                    InstrumentationVisitor(nextClassVisitor, config)
+                }
+            }
+        }
+
+        override fun isInstrumentable(classData: ClassData): Boolean {
+            val ignoresParam = parameters.get().ignores.get()
+
+            if (DebugUtil.firstInstrument) {
+                val instruLogger = Logging.getLogger(InstanaClassVisitorFactory::class.java)
+                instruLogger.debug("Ignores parameter = $ignoresParam")
+            }
+            DebugUtil.firstInstrument = false
+
+            return if (ignoresParam.any { it.matches(classData.className) }) {
+                val logger = Logging.getLogger(InstanaClassVisitorFactory::class.java)
+                logger.debug("Ignoring class ${classData.className}")
+                false
+            } else {
+                true
+            }
+        }
+    }
+}
+
+class DebugUtil {
+    companion object {
+        var firstInstrument = true
+    }
 }