+++ runtime/src/test/java/com/instana/android/core/event/worker/EventWorkerShould.kt
@@ -5,11 +5,8 @@
 
 package com.instana.android.core.event.worker
 
-import android.util.Log
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.work.*
-import androidx.work.impl.utils.SynchronousExecutor
-import androidx.work.testing.WorkManagerTestInitHelper
 import com.instana.android.BaseTest
 import com.instana.android.Instana
 import com.instana.android.InstanaShould.Companion.API_KEY
@@ -18,7 +15,7 @@ import com.instana.android.core.InstanaConfig
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
-import org.junit.Ignore
+import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
 
@@ -28,15 +25,6 @@ class EventWorkerShould : BaseTest() {
     var instantTaskExecutorRule = InstantTaskExecutorRule()
 
     init {
-        val configuration = Configuration.Builder()
-            // Set log level to Log.DEBUG to make it easier to debug
-            .setMinimumLoggingLevel(Log.DEBUG)
-            // Use a SynchronousExecutor here to make it easier to write tests
-            .setExecutor(SynchronousExecutor())
-            .build()
-
-        // Initialize WorkManager for instrumentation tests.
-        WorkManagerTestInitHelper.initializeTestWorkManager(app, configuration)
         Instana.setup(app, InstanaConfig(API_KEY, SERVER_URL))
     }
 
@@ -57,13 +45,18 @@ class EventWorkerShould : BaseTest() {
     }
 
     @Test
-    @Ignore
     fun doWorkEnqueued() {
         val directory = app.filesDir
         val request = EventWorker.createWorkRequest(Constraints.NONE, directory, 0L, "tag")
-        val workManager = WorkManager.getInstance()
+
+        val instanaWorkManager = Instana.workManager
+        Assert.assertNotNull(instanaWorkManager)
+
+        val workManager = instanaWorkManager!!.getWorkManager()
+        Assert.assertNotNull(workManager)
+
         // Enqueue and wait for result.
-        workManager.enqueue(request).result
+        workManager!!.enqueue(request).result
         // Get WorkInfo
         val workInfo = workManager.getWorkInfoById(request.id).get()
         // Assert